requirements.txt
absl-py==0.11.0
adamp==0.3.0
astunparse==1.6.3
blis==0.7.4
cachetools==4.2.0
catalogue==1.0.0
certifi==2020.12.5
chardet==3.0.4
click==7.1.2
cymem==2.0.5
dataclasses==0.6
filelock==3.0.12
flatbuffers==1.12
future==0.18.2
gast==0.3.3
google-auth==1.23.0
google-auth-oauthlib==0.4.2
google-pasta==0.2.0
grpcio==1.32.0
h5py==2.10.0
idna==2.10
joblib==1.0.1
Keras-Preprocessing==1.1.2
Markdown==3.3.3
murmurhash==1.0.5
nltk==3.5
numpy==1.19.4
oauthlib==3.1.0
opt-einsum==3.3.0
packaging==20.9
Pillow==8.0.1
plac==1.1.3
preshed==3.0.5
protobuf==3.14.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==2.4.7
PyYAML==5.4.1
regex==2020.11.13
requests==2.25.0
requests-oauthlib==1.3.0
rsa==4.6
sacremoses==0.0.43
six==1.15.0
spacy==2.3.5
srsly==1.0.5
tb-nightly==2.5.0a20201210
tensorboard-plugin-wit==1.7.0
termcolor==1.1.0
tf-estimator-nightly==2.5.0.dev2020121101
tf-nightly==2.5.0.dev20201208
thinc==7.4.5
tokenizers==0.10.1
torch==1.7.0+cu110
torchaudio==0.7.0
torchvision==0.8.1+cu110
tqdm==4.57.0
transformers==4.3.2
typing-extensions==3.7.4.3
urllib3==1.26.2
wasabi==0.8.2
Werkzeug==1.0.1
wrapt==1.12.1
